+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 23.8.0
+BUG FIXES:
+* resource/volume: fix documentation name for volume and add an example for creating on_prem volume.
+* ressource/cvo_aws, cvo_azure, cvo_gcp: remove force new from `retries`.
+
+NEW ENHANCEMENTS:
+* resource/cvo_aws and cvo_gcp: add `saas_subscription_id`.
+
+NEW FEATURES:
+* resource/cvo_gcp: support adding `firewall_tag_name_rule` and `firewall_ip_ranges`.
+
 ## 23.7.0
 NEW FEATURES:
 * resource/connector_gcp: support adding gcp keys `gcp_block_project_ssh_keys`, `gcp_serial_port_enable`, `gcp_enable_os_login` and `gcp_enable_os_login_sk` to the config.